Replace `Root` with `Element`
